[chores] Moved new estimated location logic to service class
For consistency with current architecture.
