+++ openwisp_controller/geo/estimated_location/tasks.py
@@ -6,111 +6,16 @@ from swapper import load_model
 
 from openwisp_controller.config.whois.service import WHOISService
 from openwisp_controller.geo.estimated_location.service import EstimatedLocationService
-from openwisp_controller.geo.estimated_location.utils import (
-    get_location_defaults_from_whois,
-    send_estimated_location_notification,
-)
 
 logger = logging.getLogger(__name__)
 
-
-def _handle_attach_existing_location(
-    device, device_location, ip_address, existing_device_location, whois_obj
-):
-    """
-    Helper function to:
-    1. Attach existing device's location (same last_ip) to current device, else
-    2. Update current location of device using WHOIS data; if it exists, else
-    3. Create a new estimated location for the device using WHOIS data.
-    """
-    Device = load_model("config", "Device")
-
-    current_location = device_location.location
-    attached_devices_exists = None
-    if current_location is not None:
-        attached_devices_exists = (
-            Device.objects.filter(devicelocation__location_id=current_location.pk)
-            .exclude(pk=device.pk)
-            .exists()
-        )
-    if (
-        existing_device_location
-        and existing_device_location.location != device_location.location
-    ):
-        existing_location = existing_device_location.location
-        device_location.location = existing_location
-        device_location.full_clean()
-        device_location.save()
-        logger.info(
-            f"Estimated location saved successfully for {device.pk}"
-            f" for IP: {ip_address}"
-        )
-        # We need to remove existing estimated location of the device
-        # if it is not shared
-        if attached_devices_exists is False:
-            current_location.delete()
-        send_estimated_location_notification(
-            device=device,
-            notify_type="estimated_location_updated",
-            actor=existing_location,
-            ip_address=ip_address,
-            whois=whois_obj,
-        )
-        return
-    # If existing devices with same last_ip do not have any location
-    # then we create a new location based on WHOIS data.
-    if not whois_obj or not whois_obj.coordinates:
-        logger.warning(
-            f"Coordinates not available for {device.pk} for IP: {ip_address}."
-            " Estimated location cannot be determined."
-        )
-        return
-
-    location_defaults = {
-        **get_location_defaults_from_whois(whois_obj),
-        "organization_id": device.organization_id,
-    }
-    # Create new location only if location is changed.
-    if (
-        attached_devices_exists
-        and current_location
-        and current_location.geometry == location_defaults.get("geometry")
-        and current_location.name == location_defaults.get("name")
-    ):
-        logger.debug(
-            f"Estimated location unchanged for {device.pk}"
-            f" for IP: {ip_address}, keeping existing location"
-        )
-        return
-    # create new location if no location exists for device or the estimated location
-    # of device is shared.
-    estimated_location_service = EstimatedLocationService(device)
-    estimated_location_service._create_or_update_estimated_location(
-        location_defaults, attached_devices_exists, whois_obj
-    )
-    logger.info(
-        f"Estimated location saved successfully for {device.pk} for IP: {ip_address}"
-    )
+# Celery tasks validate queued work and coordinate transactions.
+# Keep estimated-location decisions in EstimatedLocationService.
 
 
 @shared_task(name="whois_estimated_location_task")
 def manage_estimated_locations(device_pk, ip_address):
-    """
-    Creates/updates estimated location for a device based on the latitude and
-    longitude or attaches an existing location.
-    Existing location here means a location of another device whose last_ip matches
-    the given ip_address.
-    Does not alters the existing location if it is not estimated.
-
-    - If the current device has no location or location is estimated, either update
-    to an existing location; if it exists, else
-
-    - A new location is created if current device has no location, or
-    if it does; it is updated using coords from WHOIS record if it is estimated.
-
-    In case of multiple devices with same last_ip, the task will send a notification
-    to the user to resolve the conflict manually.
-    """
+    """Update a device's estimated location from its current WHOIS data."""
     Device = load_model("config", "Device")
     WHOISInfo = load_model("config", "WHOISInfo")
     normalize_ip = WHOISService.normalize_ip_address
@@ -135,59 +40,8 @@ def manage_estimated_locations(device_pk, ip_address):
             ):
                 return
             whois_obj = WHOISInfo.objects.filter(ip_address=ip_address).first()
-            _manage_estimated_location(device, ip_address, whois_obj)
+            EstimatedLocationService(device).update_from_whois(ip_address, whois_obj)
     except Device.DoesNotExist:
         logger.warning(
             f"Device {device_pk} not found, skipping manage_estimated_locations"
-        )
-
-
-def _manage_estimated_location(device, ip_address, whois_obj):
-    Device = load_model("config", "Device")
-    DeviceLocation = load_model("geo", "DeviceLocation")
-    devices_with_location = Device.objects.filter(
-        organization_id=device.organization_id,
-        _is_deactivated=False,
-        last_ip=ip_address,
-        devicelocation__location__isnull=False,
-    ).exclude(pk=device.pk)
-    # Two IDs are enough to detect whether devices behind a proxy disagree.
-    location_ids = list(
-        devices_with_location.values_list(
-            "devicelocation__location_id", flat=True
-        ).distinct()[:2]
-    )
-    # multiple devices can have same last_ip in cases like usage of proxy
-    if len(location_ids) > 1:
-        send_estimated_location_notification(
-            device=device, notify_type="estimated_location_error"
-        )
-        logger.error(
-            "Multiple devices with locations found with same "
-            f"last_ip {ip_address}. Please resolve the conflict manually."
-        )
-        return
-    # if device doesn't have a location yet, initialize a draft
-    if not (device_location := getattr(device, "devicelocation", None)):
-        device_location = DeviceLocation(content_object=device)
-    current_location = device_location.location
-    if not current_location or current_location.is_estimated:
-        existing_device_location = None
-        if location_ids:
-            # Ambiguity has been ruled out, so retrieve only the shared location.
-            existing_device_location = (
-                DeviceLocation.objects.select_related("location")
-                .filter(
-                    content_object__in=devices_with_location,
-                    location_id=location_ids[0],
-                )
-                .first()
-            )
-        _handle_attach_existing_location(
-            device, device_location, ip_address, existing_device_location, whois_obj
-        )
-    else:
-        logger.info(
-            f"Non Estimated location already set for {device.pk}. Update"
-            f" location manually as per IP: {ip_address}"
         )
